% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_image_palette.R
\name{getHexPrimaries}
\alias{getHexPrimaries}
\title{single_image_palette}
\usage{
getHexPrimaries(img, cutoff)
}
\arguments{
\item{img}{image file}

\item{cutoff}{number of colors to provide}
}
\description{
This creates a color palette from an image
}
