% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.image.R
\name{hex.image}
\alias{hex.image}
\title{hex.image}
\usage{
hex.image(x, y)
}
\arguments{
\item{x}{image file}

\item{y}{number of colors to provide}
}
\description{
This creates a color palette from an image
}
