% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImagePaletteR-package.R
\docType{package}
\name{ImagePaletteR-package}
\alias{ImagePaletteR}
\alias{ImagePaletteR-package}
\title{ImagePaletteR: create color palette from Image}
\description{
Creates a color palette from clustering a provided image to a set number of clusters.
}
\author{
\strong{Maintainer}: Gould Trevor \email{goul0109@umn.edu}

}
\keyword{internal}
