% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_colors.r
\name{show_color}
\alias{show_color}
\title{show_colors}
\usage{
show_color(
  colors,
  ncol,
  byrow = TRUE,
  label = FALSE,
  number = FALSE,
  size = 1,
  border = "black"
)
}
\description{
This function returns colors.
}
\keyword{1}
\keyword{black}
\keyword{boolean}
\keyword{border}
\keyword{byrow}
\keyword{colors}
\keyword{internal}
\keyword{label}
\keyword{list}
\keyword{ncol}
\keyword{number}
\keyword{of}
\keyword{size}
